campaigns: patch changeset IDs when migrating (#13111)
In #13001, I added some "evil" to move campaigns back to their original IDs
after everything was created. In my rush to do so, I committed one of the basic
mistakes of evil-doing: not considering the broader effects of my evil; namely,
that changesets have an effective (but unenforced) foreign key on the campaign
in their JSON campaign_ids blobs.

Obviously, the solution here is more evil, in the form of an UPDATE query that
patches the JSON en masse. You're welcome?