graphqlbackend: use FileMatchResolver methods more
This is helping decouple many uses of FileMatchResolver in our
graphqlbackend to what it is actually resolving. This commit is purely a
refactor to make it easier to change out the fields on
FileMatchResolver.

This is motivated by experiments I did with the text search layer living
completely outside of graphqlbackend. Part of that experiment was
changing FileMatchResolver to instead store the types returned by that
layer. This is a change from that series which in general is an
improvement and will make future changes clearer.