graphqlbackend: use FileMatchResolver methods more (#13165)
This is helping decouple many uses of FileMatchResolver in our
graphqlbackend to what it is actually resolving. This commit is purely a
refactor to make it easier to change out the fields on
FileMatchResolver.

This is motivated by experiments I did with the text search layer living
completely outside of graphqlbackend. Part of that experiment was
changing FileMatchResolver to instead store the types returned by that
layer. This is a change from that series which in general is an
improvement and will make future changes clearer.

We introduce two helpers:
- `path()` which returns the file's path. We often directly access JPath,
  so it makes it a pain to remove JPath. This change makes that easier to
  do.
- `appendMatches(*FileMatchResolver)` this was extracted from some and/or
  code.