search: read user settings from database once in search path
In a few places on our search path we called
decodedViewerFinalSettings. This seemed to be often used to decide to
use globbing or not. Each call in production seems to take roughly
25ms-40ms. We estimate that this will save about ~200ms on the critical
path for Sourcegraph.com search requests.

This removes most places we call it by passing down the user settings in
the searchResolver.

Co-Author: Erik Seliger