monitoring: relax cpu provisioning alert with longer duration (#13175)
Brief peaks for CPU usage are not particularly useful indicators that a service might be underprovisioned - it might even be desired behaviour, we want services to occasionally be capable of using up almost all of its capacity when needed. If a service consistently experiences high CPU usage peaks, however, that's a stronger sign that a service might require more resources.

We add a 30-minute duration to short-term provisioning to try and better capture this.