Show alerts before search results, add unit test to verify (#13229)
If GQL search results returns an alert along with results, we currently show the alert at the bottom of the results. However, because of the virtual list scrolling, this alert can get lost or be impossible to see. Moving to the top so we can in the future highlight scenarios like "did you mean".

![image](https://user-images.githubusercontent.com/206864/90921900-e5077d80-e39f-11ea-88ad-d619d386f30e.png)