dev: remove nginx
We already have Caddy running for HTTPS support. We can use just Caddy
to remove the need to have nginx in our dev environments. This does mean
we won't be using the same reverse proxy as we do in production and for
customers. However, I do not think that changes much in practice.