Fixes a regression introduced in #13271
**Before #13271**

`^revision` caused a BadCommitError, because `^revision` (with caret) could not be resolved to a commit. Because we did not handle BadCommitError `^revision` was added to the list of resolved revisions anyway, which is required downstream.

**With #13271**

`^revision` causes a BadCommitError and a warning is returned to the user

**After this PR**

we strip the prefix `^` and validate the revision. If for `^<revision>`, <revision> is a valid revision, no warning is returned.

### Positive side effect:

- revisions with prefix `^` are not validated and reported missing just like other revisions
- I covered the logic with tests so introducing further regressions should be harder