monitoring: add fs_io_operations alert for gitserver, zoekt (#13062)
Add monitoring for gitserver and zoekt-* services read and write operations. We use a sum of reads and writes rates over 1h to get a picture of sustained IO usage, and set the alert very high since it's currently unclear how we might want to alert on this at the moment.