graphqlbackend: only trace graphql queries (#12977)
Previously we would trace every graphql field as well. This lead to very
large traces when returning search results. This is due to every results
having multiple fields associated with it. The interesting information
contained in the traces is our requests to databases/etc, while the
field responsible for the request was largely noise. For large results
this lead to it being impossible to load the Jaeger UI.

For example a search request on my localhost went from 161 spans to 37.

Co-Author: Stefan Hengl <stefan@sourcegraph.com>