Define diff stats next to diff in tests
I that we had a few of these diff.Stats lying around, refering to
different diffs, so I put them next to the diffs themselves.

In the case of MockChangesetSyncState I actually defined it on the
`state` itself, which I'm not 100% sure about yet, but I think it's
pretty neat to be able to access the properties of the mocked state
through the `state` itself.