Make "don't build ctags image" check less agressive (#13366)
Turns out my previous PR #13218 was a regression.

We set `CTAGS_COMMAND` to `cmd/symbols/universal-ctags-dev` in
`./dev/start.sh` by default. That means if you've never built the Docker
image, we'd never build it with the old code.

This changes the fix so that we only don't build the image if you have a
custom executable specified.