Fix race condition in campaigns.Service tests
This mock was set up by this subtest and the one for
`CloseOpenChangesets`, but only used by the latter.

Since setting up and tearing down the mock modifies global state, we
reduce the second setup (since we don't need it) to get rid of the race
condition.

And in #13367 the CloseOpenChangesets method will be removed.