search: limit queries with before/after as well
On sourcegraph.com a diff query with no repo filter and an after filter
will be run on 200k repos. With saved searches this has lead to periodic
issues (~15min) of gitserver IO being maxed out and frontend's goroutine
count spiking to 200k+. This has likely contributed to a lot of
instability. This only started becoming a problem when we bumped the
count of default repos from 10k to 200k.

The limit of 1000 is a balance between system stability and trying to
prevent breaking saved searches. I'm happy to make it larger if needs be
for saved searches at customer sites.