search: configurable limits on commit/diff search (#13439)
We recently introduced a limit of 10k on commit/diff searches with
"after:" or "before:" clauses to resolve a site issue. However, this
could break existing customer's saved searches. So we introduce a site
configuration for this limit. Along with this change we also made it
possible to configure the previous limit as well as a consistent name
for maxReposToSearch.