schema: fix incorrectly nested prop in settings (#13472)
This was introduced in #8059, but appears to have been placed in the wrong part
of the schema document. Since this setting is only used by the web frontend,
and it's accessed from a object by its string key, there should be no
functional change, besides the slim chance that a user has a non-boolean value
at present, and the user settings textarea now autocompleting for
`search.hideSuggestions`.