authz: correctly marshal Perms for logging (#13522)
In go1.15 the authz tests fail with the following: conversion from
Perms (uint32) to string yields a string of one rune, not a string of
digits (did you mean fmt.Sprint(x)?)

Instead of directly calling "Perm.String()" (or indirectly via
"fmt.Sprint") we use trace.Stringer which is a lazy logger. It will only
call the function if we actually consume the log somewhere (so will
avoid the work if we are not tracing that request).