search: combine logging for slow searches and honeycomb
This is the same pattern we use in gitserver. We can use the event we
build up for honeycomb and log it to log15.

We also make the decision to do slow query logging a bit cleaner by
extracting logSlowSearchesThreshold into a function. Additionally we no
longer log the username. Rather the Actor UID and if the actor is
internal. This avoids a DB lookup and will allow us to differentiate
between internal and unauthenticated users.