gitserver: create reposstats endpoint (#13760)
We want to compute the total number of bytes for git repos on a
gitserver. This commit adds a new endpoint to gitserver for
ReposStats. ReposStats currently has a field for total number of bytes
of all git repos. In future we could also add other relevant statistics.

We compute the total size of the repositories as part of our janitor
jobs. We do this because this can be slow to compute. The janitor jobs
are useful infrastructure which periodically run over all repositories
on a gitserver.

There are no callers of the client yet. We intend to use this in site
pings.

Co-authored-by: ᴜɴᴋɴᴡᴏɴ <joe@sourcegraph.com>