CampaignDetailsPage: tab bar horizontal line takes up full width
This makes it easier to visually distinguish that everything below the tab bar is an entirely separate content area from what's above. Previously the horizontal line only extended the width of the container, which does not make sense visually IMO (the tab panel area needs to be an enclosed shape, bounded by borders or browser window boundaries on all 4 sides).

All other campaigns pages remain the same. The diff is deceptively large because a lot of code had to be un-indented (to remove a wrapping `div.container`) or indented (to add a wrapping `div.container`), which is a big diff but an insignificant change.