Update commit when author email/name changed (#13819)
I noticed that the reconciler didn't create and push a new commit when I
changed the author in my campaign spec and applied it.

This fixes that issue. The code is simply an addition to what was there
before and doesn't change anything.

I want to investigate how/whether I can make CompareChangesetSpecs
nicer, but I'll leave that for a follow-up.