Misc campaigns improvements (#13783)
* capitalize YAML in "Download YAML" button

This is more common and looks more correct/intentional.

* CampaignSpecTab: use full width

Previously, this was the only tab whose content was centered. It looked inconsistent. Now it takes up the full width.

* fix "Download YAML" filename and tooltip

- The filename convention we've defined is `foo.campaign.yaml` not `foo.spec.yaml`.
- Mention the actual filename in the tooltip, not `campaign-spec.yaml`.

* CampaignSpecTab: more precise explanation of how the spec was used

- Differentiates between created and updated. Previously, it said "created", which is technically incorrect if the campaign had been updated with a different spec after creation.
- Adds a snapshot test.

* tighten up CampaignInfoByline

Reduces duplication and makes this shorter. I also prefer not having sentence-case prose text right below the title, because some people will be confused and visually scan it as the campaign description (which it isn't).

Before: "Created 3 days ago by alice | Updated 3 days ago by bob"

After: "alice 3 days ago" (if no updates yet), "alice 3 days ago • [bob] updated 2 days ago" (where "[bob]" is omitted if the user who updated it is the same as the user who created it)

* CampaignDescription: take up full width

Previously, the campaign description was indented. This meant (IMO) that the left edge of the campaign page was ragged/uneven.

* hide UserAreaHeader on campaigns pages

The UserAreaHeader (which shows the username and the Profile/Settings/Saved searches/Event log/Campaigns tabs) is not very useful on campaigns pages, and it takes a lot of vertical space.

* CampaignDetailsPage: tab bar horizontal line takes up full width

This makes it easier to visually distinguish that everything below the tab bar is an entirely separate content area from what's above. Previously the horizontal line only extended the width of the container, which does not make sense visually IMO (the tab panel area needs to be an enclosed shape, bounded by borders or browser window boundaries on all 4 sides).

All other campaigns pages remain the same. The diff is deceptively large because a lot of code had to be un-indented (to remove a wrapping `div.container`) or indented (to add a wrapping `div.container`), which is a big diff but an insignificant change.

* CampaignCloseAlert: remove shadow effect

This shadow effect is not used anywhere else in our product and is not in our style guide (AFAICT).

* make campaigns icon flush with the left edge of other content

mdi-react icons all contain built-in spacing of (generally) 3px on all sides (in the 24x24 px reference size). When the icon is used in the title of a page, such as "(Campaigns icon) alice / campaign-name", this causes the campaigns icon to be indented more than the content below it, which looks bad.