symbols: remove language list
By specifying no languages we get all languages (the
default). Previously we specified this list to ensure we had the correct
version of ctags running. Now that we pass in the custom language
regexes we know we are always running the correct version.

Additionally we add a test using graphql which would fail if we do this
incorrectly.