indent and highlight JSON campaign specs
A campaign spec is a YAML document. YAML is a superset of JSON. If a campaign spec is pure-JSON, then it's nice to indent it (in case it's all on one line, for example) and highlight it using the JSON highlighter.

The button is still "Download YAML" and the filename is still "whatever.yaml", but that's OK because YAML is a superset of JSON, so those are correct.