web: initial search stream PoC (#14038)
This is an initial implementation of streaming search results from the
backend. It is based on our proof of concept backend layer. From here we
can interatively improve and change the protocol to suit the frontend's
needs. This is not production ready and as such is hidden behind a
feature flag.

This implementation is a bit hacky. The GraphQL types are quite heavily
baked into our components. So this implementation converts the stream
into the graphql types.

Co-authored-by: Loïc Guychard <loic@sourcegraph.com>