symbols: do not rely on ctags.d (#14037)
We have regressed on this before by having .ctags.d in the incorrect
place in a container. Instead we can just pass the values in as command
line flags (which is what the files are). We get the values by
introducing a go generate step which embeds the flags into a file as a
`[]string`.

Additionally this makes it easier to use a our own CTAGS_COMMAND in
development.

My intention is to pull this code out into a shared lib and get Zoekt to
use the same library. This should remove inconsistencies between the two
environments as well as remove extra requirements on how we build
containers.