symbols: remove .ctags.d (#14065)
In my previous symbols change I didn't properly test runnings ctags
with it finding the ".ctags.d". With both it leads to flags being
defined twice. Instead we need to remove including .ctags.d in both our
development environment and in the containers we build.

This commit removes .ctags.d from all places it is mentioned. Instead we
move the files into a directory "ctags.d" that universal-ctags won't
look at. Note: the previous change manually specified the flags that are
in ctags.d, so we won't regress on symbol finding for the additional
languages.