simplify extension API context impl
The getComputedContextProperty function is called to provide the values for extension context keys such as `resource.uri` (for extensions to define actions that, when invoked, call a function with the URI of the currently visible document).

Now, instead of getComputedContextProperty parsing the requested key (such as `resource.uri`) and running custom code to return each key's value, it constructs a JavaScript object with all possible keys and just indexes into that object.

This new impl is simpler. It's also a milestone along the way to further simplification of this (to remove the need to call this function for each property altogether, instead allowing callers to just index into the pre-computed comprehensive context object themselves).