finish simplification of extension API context impl
The getComputedContextProperty is renamed to computeContext. It now always returns the entire computed context data as a map, instead of computing the value per-key. This is a simplification because it enforces that there must be no special per-key computation; the context can be treated as a map instead of as an interface with an arbitrary `{get(key:string): any}` function.