remove search scope pages (/search/scope/:id) (#14127)
The search scope page was a page at `/search/scope/:id` (where `:id` is the ID of a search scope in settings) that displayed the name and description of the scope, plus a search bar that searched within the scope.

It was never documented (other than in the settings JSON Schema), usage was very low on Sourcegraph.com (not surprisingly, because it was disabled in 2018), and we only used it at in-person conference booths more than a year ago. We don't have any data on whether customers use this feature, but given these factors, we can assume that they do not (and if we find out otherwise before the release, we can plan accordingly).