show BrowserExtensionToast after more days of activity (#14261)
Currently, the "Get Sourcegraph on GitHub" toast shows up on the first day of being active. That means that a brand new user who just visits Sourcegraph for the first time will see it. This is not effective because it's (1) annoying and (2) before they have had time to (hopefully) enjoy using Sourcegraph.