search: send dynamic filters in stream api (#14238)
This is a proposed API for sending dynamic filters via the streaming API. When
a new "filters" event is sent, it replaces the existing list.

The proposed API is exactly the same as our graphql type. I'm unsure if we can
make this smaller at all, nor if it is worth doing it since this list is
usually small.