web/search: scan over ISearchResults in stream (#14309)
This simplifies the observable chain so we directly update the final
ISearchResults in scan for each stream event. The main benefit of this
is as we add more event types we don't need to grow the list of event
buffers.

We also use switch on the event type so that typescript checks
exhaustiveness.