repo-updater: Sync single external service at a time using workerutil (#14305)
This PR makes two main changes:

We sync each external service independently
The syncing process now happens using the workerutil package
Syncer.Run now enqueues external services into a job queue that are due to be synced every minute.

These jobs are picked by up to three sync workers concurrently each worker syncs one service using the new Syncer.SyncExternalService method.

We rely on database triggers to mark repos as deleted once they no longer exist in any external service. This is based on the state of the external_service_repos table which tracks the relationship between repos and external services.

Triggers also exist that remove rows from external_service_repos when a repo or external service is deleted or soft deleted.

We've tested syncing up to 500k repos successfully.

* repo-updater: Sync single external service at a time using workerutil

Tests are still failing

* repoupdater: Fix server_test

* testSyncerSyncWithErrors uses SyncExternalService

Also removes a test case that tests a sync of muktiple external services
as we no longer support that.

* Add EnqueueJobs tests

* Removed old references to syncer.Sync

Code now compiles but a lot of tests still fail.

Added a cleanup function that needs to be called when we stop
a worker to unregister prometheus metrics so that it doesn't panic when we
start another worker in a subsequent test.

* Fix repo-updater tests

* Delete orphaned repos

* Avoid constraints by changing upsert order

* Fix minSyncInterval

* WIP test sync

* Add a test that syncs multiple external services in sequence

* WIP test for orphaned repos

* Only delete orphaned repos

* repo-updater: Remove stale jobs on startup

If a job is processing when repo-updater dies that job would never
complete. We would then not requeue the associated external service for
syncing. To be safe we delete non locked processing rows on startup.

* Typo

* db: Mark next_synced_at on service save

This will cause a sync to be triggered ASAP for the saved external
service.

On save, we trigger a call to repo-updater which causes us to
enqueue any pending sync jobs. As we've just upated next_sync_at the job
for the newly saved repo will be queued.

* Rename trigger

We no longer trigger a full sync but instead trigger enqueueing pending
sync jobs.

* Remove orphaned repo code

We now have a trigger that does the same thing

* Improve orphaned repo test

* Test external service deletion

* Remove orphaned repo code from syncSubset

* Remove leftover debug line

* Fix deleting of repos when called via syncSubset

* Improve EnqueueSyncJobs docstring

Co-authored-by: ᴜɴᴋɴᴡᴏɴ <joe@sourcegraph.com>

* Improve error message

Co-authored-by: ᴜɴᴋɴᴡᴏɴ <joe@sourcegraph.com>

* Check rows.Err() after iteration

* Change log line to debug

Co-authored-by: ᴜɴᴋɴᴡᴏɴ <joe@sourcegraph.com>

* Remove leftover debug line

* Add stalled jobs resetter

* Pass prometheus register to the SyncWorker

* Create ResetterMetrics inline

* Remove unused field

* Fix docstring

* Typo

* Missing period

* Typo

* Add failing test

* Resolve name conflicts deterministically

* Cleanup name conflicts resolution

* Resolve name conflicts during renames

* Don't wrap nil error

* Don't queue jobs for Phabricator services

* Handle large name lists in StoreListReposArgs.Names

* Trigger rebuild

* Remove stale job cleanup code

We use the workerutil resetter now instead

* Add failing test

* Fix predicate on sourcegraphcom mode

* Don't sync deleted external services

* When in cloud mode DON'T sync admin added external services

* Fix slow query

* Account for sourced repo Sources map always containing one source

Ensure that the Sources map on the newly sourced repo also contains any
other sources we know about from the DB so that are diff functions can
operate correctly.

The repo from the source can by definition only ever know about on
source.

* Use URI when adding stored sources

Before we used id which will always be 0 from a sourced repo as id is
from our database.

* Fix failing test

The test was no longer valid as we now ensure that sourced repos also
include any other sources we know about in our DB.

* Add comment

* Add some test helpers

* Disallow external services owned by a user to sync private code

* User more test helpers

* Temp logging

* More temp logging

* Remove temp logging

* Do not store github ViewPersmissions in our database

The value changes depending on which token was used to fetch the repo.

* Store prometheus Registerer in Syncer and record sync backoff

* Record external service id with sync duration

* Include external service id and family in sync metrics

* Simplify syncSubset signature (#13911)

* Change the signature of SyncSubset to expect only one repo

* Rename syncSubset to syncSingleRepo

* Renamed SyncSingleRepo to SyncRepo

* Streaming inserter should not insert private repos owned by a user

* Rename store to tx in SyncExternalService

* Streaming inserter should not use a transaction

It should use a store created outside of the workerutil package so that
repos added are available asap and not delayed until the entire sync has
completed.

* Added failing test testConflictingSyncers

* Make test more deterministic

* Use separate Read Commited transaction in the handler

* repo-updater: Use s.Logger where appropriate  (#14095)

* repo-updater: Use s.Logger where appropriate

The syncer has it's own logger with should be used when logging
from within any methods on the syncer struct.

Also updated a couple of places where we were using the correct logger
but not performing a nil check.

* Demote log

* Return early when UpsertSources has no work

* Use IsolationLevel option

* Make the number of workers configurable

* Update comment

* Typo

* Fix comment

* Typo

* Use hardcoded number of workers

* Consistent spacing

* Disable git update scheduling on cloud

* Remove unused context

Co-authored-by: Asdine El Hrychy <asdine.elhrychy@gmail.com>
Co-authored-by: ᴜɴᴋɴᴡᴏɴ <joe@sourcegraph.com>