graphqlbackend: remove unused types
These are types in our graphql API which are mentioned in any other
type/mutation/etc. This commit removes

- DefaultSettings
- DeploymentConfiguration
- Diff
- DiffSearchResult
- ExtensionRegistryPublishExtensionResult
- HiddenChangesetSpec
- HiddenExternalChangeset
- VirtualFile

It doesn't yet remove the relevant resolvers for those types. That will
be done in a later commit once we confirm these can be removed.

Context: I came across DiffSearchResult and noticed it was unused. I
then cooked up this little script to find all the unused types:

  (
      for t in $(awk '/^(type|enum|interface|union)/ { print $2 }' cmd/frontend/graphqlbackend/schema.graphql); do
          echo -n "$t"
          egrep "\b$t\b" cmd/frontend/graphqlbackend/schema.graphql | wc -l
      done
  ) | grep ' 1$' | sort | awk '{print $1}'