Test ping event structure equality as a whole.
Since the event log insertion takes a `db.Event` and the return from the queries are e.g. `types.Event`, this introduces an `expected` test variable to do the comparison. Some fields, such as the timestamp or application version, are automatically filled in by the `db` package so I've explicitly set them here in the same way.

ref https://github.com/sourcegraph/sourcegraph/pull/14370#discussion_r498786735