allow cookiebot to scan sourcegraph.com by disabling redirection (#14381)
Currently, Cookiebot is not scanning sourcegraph.com at all, and instead is being redirected to about.sourcegraph.com. We need to allow Cookiebot to scan the main website so we can catalog which cookies we use. [Cookiebot recommends doing this by checking for the string `Cookiebot` in the user agent.](https://support.cookiebot.com/hc/en-us/articles/360003824153)