fix result type of Mutation.publishExtension (#14327)
It was a mistake that the createExtension result type was used for this as well. No callers (that I'm aware of, or that are likely to exist otherwise) rely on the previous type name, and the types currently have the same shape, so the risk of breakage is extremely low.