Simplify getProviders()
getProviders() was originally passed Observables for patternType, globbing, to avoid re-registering providers if those values change. In practice, this was premature / unneeded optimizations, as those values will not change often.

In this commit, I refactor getProviders() to take an options object for those values, and re-register providers when they change. This makes it obvious how to pass down new, similar values (like interpretComments) to getProviders().

There is a Best Solution here which is to convert MonacoQueryInput to a function component, so that we can simply handle registering providers in a `useEffect()` block (like we do in SearchConsole), instead of having to do the strange `new Observable()` dance here. I started on this, but have a few nasty bugs to iron out today, so here is an incremental solution.