search: indexed config API supports multiple repo query arguments (#14558)
This converts the API into a batch API by writing a JSON blob per "repo"
query parameter in the order they are specified.

Each zoekt indexserver currently queries this endpoint with a hardcoded
concurrency of 32. This can cause many many HTTP requests, which on a
linux machine without tuning can lead to socket errors. By converting
this into a batch API we reduce the number of connections as well as
moving the responsbility of concurrency to the frontend which has more
knowledge of load/etc.

We also remove backwards compatibility support where repo wasn't
specified. Repo has been specified for multiple releases now.