repo-updater: Add critical alerts (#14530)
* repo-updater: Add critical alerts

The number of user added repos is greater than 90% of our hard limit of
200k.
Number of external services added is great than 20k
No external services have synced in more than 8 hours. By default we
backoff for up to 8 hours, so anything higher than this indicates a
problem.

* syncer_sync_start should be a rate

Alert when we've performed <= 1 syncs for 8 hours.

Ideally we want this to be <= 0 but our monitoring package doesn't allow
this. Therefore we need to make this a warning as it'll fire for
instances that only have 1 external service defined.

* Fill in possible solutions and remove unnecessary panel

* Adjust rate

* Update monitoring/repo_updater.go

Co-authored-by: Tomás Senart <tomas@sourcegraph.com>

* Update monitoring/repo_updater.go

Co-authored-by: Robert Lin <robert@bobheadxi.dev>

* Update possible error descriptions

* Fix alert

* Switch alert to critical

Co-authored-by: Tomás Senart <tomas@sourcegraph.com>
Co-authored-by: Robert Lin <robert@bobheadxi.dev>