fix regression in extension API to allow dynamic context key access (#14587)
In https://github.com/sourcegraph/sourcegraph/pull/14088 I simplified how the context was computed. But I broke support for <code>get(context, `some-dynamically-\${interpolated}-string`)</code>; I accidentally removed the ability to reference `context` and index into the whole map, and there was no test case for that.

The fix here is to treat `context` as a special global, like `true` and `false`. An alternative fix would be to make the context map itself contain a key named `context` that was self-referential, but that leads to a cyclic data structure, which sounds like a bad idea.

fix #14283